% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_small}
\alias{pbmc_small}
\title{pbmc_small}
\format{
A Seurat object.
}
\usage{
data(pbmc_small)
}
\description{
A small Seurat dadaset of single cell RNA sequencing data
}
\keyword{datasets}
